Split history persistence adapters
