+++ apps/server/tests/hygiene/test_history_db_mixins.py
@@ -1,8 +1,17 @@
-"""Guardrails for the intentional ``HistoryDB`` mixin composition."""
+"""Guardrails for the split history persistence composition."""
 
 from __future__ import annotations
 
-from vibesensor.adapters.persistence.history_db import HistoryDB
+from pathlib import Path
+
+from vibesensor.adapters.persistence.history_db import (
+    ClientNameRepository,
+    HistoryDB,
+    RunHistoryRepository,
+    SettingsSnapshotRepository,
+    SQLiteHistoryEngine,
+    create_history_persistence_adapters,
+)
 from vibesensor.adapters.persistence.history_db._queries import _HistoryDBQueryMixin
 from vibesensor.adapters.persistence.history_db._run_lifecycle import _HistoryDBRunLifecycleMixin
 from vibesensor.adapters.persistence.history_db._sample_io import _HistoryDBSampleIOMixin
@@ -16,7 +25,7 @@ def _public_method_names(cls: type[object]) -> set[str]:
     }
 
 
-def test_history_db_mixins_expose_disjoint_public_methods() -> None:
+def test_run_history_repository_mixins_expose_disjoint_public_methods() -> None:
     mixins = (
         _HistoryDBRunLifecycleMixin,
         _HistoryDBSampleIOMixin,
@@ -29,20 +38,29 @@ def test_history_db_mixins_expose_disjoint_public_methods() -> None:
             owners_by_name.setdefault(name, []).append(mixin.__name__)
 
     overlaps = {name: owners for name, owners in owners_by_name.items() if len(owners) > 1}
-    assert overlaps == {}, f"HistoryDB mixins must keep disjoint public APIs: {overlaps}"
+    assert overlaps == {}, f"run-history mixins must keep disjoint public APIs: {overlaps}"
 
 
-def test_history_db_wrapper_owns_shared_cursor_contract() -> None:
+def test_history_db_wrapper_is_no_longer_the_primary_mixin_stack() -> None:
     history_db_members = set(HistoryDB.__dict__)
 
-    assert {"_cursor", "_cursor_connection"} <= history_db_members
-    assert "write_transaction_cursor" in history_db_members
+    assert not issubclass(HistoryDB, _HistoryDBRunLifecycleMixin)
+    assert not issubclass(HistoryDB, _HistoryDBSampleIOMixin)
+    assert not issubclass(HistoryDB, _HistoryDBQueryMixin)
+    assert "__getattr__" in history_db_members
+    assert {"_cursor", "_cursor_connection", "write_transaction_cursor"} <= history_db_members
+
+
+def test_history_persistence_factory_returns_split_adapters(tmp_path: Path) -> None:
+    adapters = create_history_persistence_adapters(tmp_path / "history.db")
+    try:
+        assert isinstance(adapters.lifecycle, SQLiteHistoryEngine)
+        assert isinstance(adapters.run_repository, RunHistoryRepository)
+        assert isinstance(adapters.settings_snapshot_repository, SettingsSnapshotRepository)
+        assert isinstance(adapters.client_name_repository, ClientNameRepository)
+    finally:
+        adapters.lifecycle.close()
 
     assert "_cursor_connection" not in _HistoryDBRunLifecycleMixin.__dict__
     assert "write_transaction_cursor" not in _HistoryDBSampleIOMixin.__dict__
     assert "write_transaction_cursor" not in _HistoryDBQueryMixin.__dict__
-
-    for mixin in (_HistoryDBSampleIOMixin, _HistoryDBQueryMixin):
-        assert "_cursor_connection" not in mixin.__dict__, (
-            "sample/query mixins must rely on the wrapper-owned cursor-selection logic"
-        )
