+++ apps/server/vibesensor/app/runtime_state.py
@@ -22,10 +22,10 @@ from vibesensor.use_cases.updates.manager import UpdateManager
 
 if TYPE_CHECKING:
     from vibesensor.adapters.gps.gps_speed import GPSSpeedMonitor
-    from vibesensor.adapters.persistence.history_db import HistoryDB
     from vibesensor.adapters.udp.udp_control_tx import UDPControlPlane
     from vibesensor.adapters.websocket.hub import WebSocketHub
     from vibesensor.app.settings import AppConfig
+    from vibesensor.infra.runtime.lifecycle import LifecycleHistoryDb
 
 
 @dataclass(slots=True)
@@ -39,7 +39,7 @@ class RuntimeState:
     worker_pool: WorkerPool
     settings_store: SettingsReader
     gps_monitor: GPSSpeedMonitor
-    history_db: HistoryDB
+    history_db: LifecycleHistoryDb
     processing_loop_state: ProcessingLoopState
     health_state: RuntimeHealthState
     processing_loop: ProcessingLoop
